+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Make `Board::new(p, cp)` public and fix RTIC example
 - Fix display-nonblocking example
 - Fix timer for LED-display (GreyscaleImage with a empty Row did not work)
 
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Update dependencies nrf51-hal and nrf52833-hal to 0.14.0
 - Added TEMP field to board
+- Fixed Issue where columns 2,3 and 4 of the nonblocking display were swapped
 
 ## [0.11.0] - 2021-09-13